adjust and simply refactor
