+++ tee-worker/litentry/core/assertion-build/Cargo.toml
@@ -52,6 +52,7 @@ sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "polkad
 lc-data-providers = { path = "../data-providers", default-features = false }
 lc-stf-task-sender = { path = "../stf-task/sender", default-features = false }
 litentry-primitives = { path = "../../primitives", default-features = false }
+parachain-core-primitives = { package = "core-primitives", path = "../../../../primitives/core", default-features = false }
 
 [features]
 default = ["std"]
