+++ tee-worker/litentry/core/stf-task/receiver/Cargo.toml
@@ -57,6 +57,8 @@ lc-identity-verification = { path = "../../identity-verification", default-featu
 lc-stf-task-sender = { path = "../sender", default-features = false }
 litentry-primitives = { path = "../../../primitives", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.36", default-features = false }
+parachain-core-primitives = { package = "core-primitives", path = "../../../../../primitives/core", default-features = false }
+
 
 [features]
 default = ["std"]
